fix(local-k8s): remove hardcoded ingress class
