+++ garden-service/src/plugins/kubernetes/local/local.ts
@@ -6,155 +6,19 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import * as execa from "execa"
-import { safeLoad } from "js-yaml"
-import * as Joi from "joi"
-import { join } from "path"
-import { GardenPlugin, PluginFactoryParams } from "../../../types/plugin/plugin"
-import {
-  gardenPlugin as k8sPlugin,
-  KubernetesBaseConfig,
-  kubernetesConfigBase,
-} from "../kubernetes"
-import { readFile } from "fs-extra"
-import { homedir } from "os"
+import { GardenPlugin } from "../../../types/plugin/plugin"
+import { gardenPlugin as k8sPlugin } from "../kubernetes"
 import { getLocalEnvironmentStatus, prepareLocalEnvironment } from "../init"
-import { ConfigureProviderParams } from "../../../types/plugin/params"
-
-// TODO: split this into separate plugins to handle Docker for Mac and Minikube
-
-// note: this is in order of preference, in case neither is set as the current kubectl context
-// and none is explicitly configured in the garden.yml
-const supportedContexts = ["docker-for-desktop", "minikube"]
-const kubeConfigPath = join(homedir(), ".kube", "config")
-
-async function getKubeConfig(): Promise<any> {
-  try {
-    return safeLoad((await readFile(kubeConfigPath)).toString())
-  } catch {
-    return {}
-  }
-}
-
-/**
- * Automatically set docker environment variables for minikube
- * TODO: it would be better to explicitly provide those to docker instead of using process.env
- */
-async function setMinikubeDockerEnv() {
-  const minikubeEnv = await execa.stdout("minikube", ["docker-env", "--shell=bash"])
-  for (const line of minikubeEnv.split("\n")) {
-    const matched = line.match(/^export (\w+)="(.+)"$/)
-    if (matched) {
-      process.env[matched[1]] = matched[2]
-    }
-  }
-}
-
-export interface LocalKubernetesConfig extends KubernetesBaseConfig {
-  _system?: Symbol
-  setupIngressController: string | boolean | null
-}
-
-export const configSchema = kubernetesConfigBase
-  .keys({
-    namespace: Joi.string()
-      .default(undefined, "<project name>")
-      .description(
-        "Specify which namespace to deploy services to (defaults to the project name). " +
-        "Note that the framework generates other namespaces as well with this name as a prefix.",
-      ),
-    setupIngressController: Joi.string()
-      .allow("nginx", false, null)
-      .default("nginx")
-      .description("Set this to null or false to skip installing/enabling the `nginx` ingress controller."),
-    _system: Joi.any().meta({ internal: true }),
-  })
-  .description("The provider configuration for the local-kubernetes plugin.")
+import { configureProvider, configSchema } from "./config"
 
 export const name = "local-kubernetes"
 
-export function gardenPlugin({ projectName, log }: PluginFactoryParams): GardenPlugin {
+export function gardenPlugin(): GardenPlugin {
   const plugin = k8sPlugin()
 
   plugin.configSchema = configSchema
 
-  plugin.actions!.configureProvider = async ({ config }: ConfigureProviderParams<LocalKubernetesConfig>) => {
-    let context = config.context
-    let defaultHostname = config.defaultHostname
-    let setupIngressController = config.setupIngressController
-
-    if (!context) {
-      // automatically detect supported kubectl context if not explicitly configured
-      const kubeConfig = await getKubeConfig()
-      const currentContext = kubeConfig["current-context"]
-
-      if (currentContext && supportedContexts.includes(currentContext)) {
-        // prefer current context if set and supported
-        context = currentContext
-        log.debug({ section: name, msg: `Using current context: ${context}` })
-      } else if (kubeConfig.contexts) {
-        const availableContexts = kubeConfig.contexts.map(c => c.name)
-
-        for (const supportedContext of supportedContexts) {
-          if (availableContexts.includes(supportedContext)) {
-            context = supportedContext
-            log.debug({ section: name, msg: `Using detected context: ${context}` })
-            break
-          }
-        }
-      }
-    }
-
-    if (!context) {
-      context = supportedContexts[0]
-      log.debug({ section: name, msg: `No kubectl context auto-detected, using default: ${context}` })
-    }
-
-    if (context === "minikube") {
-      await execa("minikube", ["config", "set", "WantUpdateNotification", "false"])
-
-      if (!defaultHostname) {
-        // use the nip.io service to give a hostname to the instance, if none is explicitly configured
-        const minikubeIp = await execa.stdout("minikube", ["ip"])
-        defaultHostname = `${projectName}.${minikubeIp}.nip.io`
-      }
-
-      if (config.setupIngressController === "nginx") {
-        log.silly("Using minikube's ingress addon")
-        await execa("minikube", ["addons", "enable", "ingress"])
-        // make sure the prepare handler doesn't also set up the ingress controller
-        setupIngressController = false
-      }
-
-      await setMinikubeDockerEnv()
-
-    } else {
-      if (!defaultHostname) {
-        defaultHostname = `${projectName}.local.app.garden`
-      }
-    }
-
-    config = {
-      name: config.name,
-      context,
-      defaultHostname,
-      deploymentRegistry: {
-        hostname: "foo.garden",   // this is not used by this plugin, but required by the base plugin
-        namespace: "_",
-      },
-      forceSsl: false,
-      imagePullSecrets: config.imagePullSecrets,
-      ingressHttpPort: 80,
-      ingressHttpsPort: 443,
-      ingressClass: "nginx",
-      namespace: config.namespace || projectName,
-      setupIngressController,
-      tlsCertificates: config.tlsCertificates,
-      _system: config._system,
-    }
-
-    return { name: config.name, config }
-  }
+  plugin.actions!.configureProvider = configureProvider
 
   // override the environment configuration steps
   plugin.actions!.getEnvironmentStatus = getLocalEnvironmentStatus
